ь-3

+++ main1.py
@@ -1,13 +1,28 @@
-from flask import Flask, redirect, render_template, request, url_for, current_app
+from flask import current_app, Flask, redirect, render_template
+from flask import request, url_for, session
 import booksdb
 import storage
 import secrets
+import json
+import os
+from urllib.parse import urlparse
+import oauth
+import translate
+import profiledb
 
 app = Flask(__name__)
 app.config.update(
     SECRET_KEY=secrets.get_secret('flask-secret-key'),
     MAX_CONTENT_LENGTH=8 * 1024 * 1024,
     ALLOWED_EXTENSIONS=set(['png', 'jpg', 'jpeg', 'gif']),
+    CLIENT_SECRETS=json.loads(secrets.get_secret('bookshelf-client-secrets')),
+    SCOPES=[
+        'openid',
+        'https://www.googleapis.com/auth/contacts.readonly',
+        'https://www.googleapis.com/auth/userinfo.email',
+        'https://www.googleapis.com/auth/userinfo.profile',
+    ],
+    EXTERNAL_HOST_URL=os.getenv('EXTERNAL_HOST_URL'),
 )
 app.debug = True
 app.testing = False
@@ -31,6 +46,113 @@ def upload_image_file(img):
 
     return public_url
 
+
+def logout_session():
+    """
+    Clears known session items.
+    """
+    session.pop('credentials', None)
+    session.pop('user', None)
+    session.pop('state', None)
+    session.pop('error_message', None)
+    session.pop('login_return', None)
+    return
+
+
+def external_url(url):
+    """
+    Cloud Shell routes https://8080-***/ to localhost over http
+    This function replaces the localhost host with the configured scheme + hostname
+    """
+    external_host_url = current_app.config['EXTERNAL_HOST_URL']
+    if external_host_url is None:
+        # force https
+        if url.startswith('http://'):
+            url = f"https://{url[7:]}"
+        return url
+
+    # replace the scheme and hostname with the external host URL
+    parsed_url = urlparse(url)
+    replace_string = f"{parsed_url.scheme}://{parsed_url.netloc}"
+    new_url = f"{external_host_url}{url[len(replace_string):]}"
+    return new_url
+
+
+@app.route('/error')
+def error():
+    """
+    Display an error.
+    """
+    if "error_message" not in session:
+        return redirect(url_for('.list'))
+
+    # render error
+    return render_template('error.html', error_message=session.pop('error_message', None))
+
+
+@app.route("/login")
+def login():
+    """
+    Login if not already logged in.
+    """
+    if not "credentials" in session:
+        # need to log in
+
+        current_app.logger.info('logging in')
+
+        # get authorization URL
+        authorization_url, state = oauth.authorize(
+            callback_uri=external_url(url_for('oauth2callback', _external=True)),
+            client_config=current_app.config['CLIENT_SECRETS'],
+            scopes=current_app.config['SCOPES'])
+
+        current_app.logger.info(f"authorization_url={authorization_url}")
+
+        # save state for verification on callback
+        session['state'] = state
+
+        return redirect(authorization_url)
+
+    # already logged in
+    return redirect(session.pop('login_return', url_for('.list')))
+
+
+@app.route("/oauth2callback")
+def oauth2callback():
+    """
+    Callback destination during OAuth process.
+    """
+    # check for error, probably access denied by user
+    error = request.args.get('error', None)
+    if error:
+        session['error_message'] = f"{error}"
+        return redirect(url_for('.error'))
+
+    # handle the OAuth2 callback
+    credentials, user_info = oauth.handle_callback(
+        callback_uri=external_url(url_for('oauth2callback', _external=True)),
+        client_config=current_app.config['CLIENT_SECRETS'],
+        scopes=current_app.config['SCOPES'],
+        request_url=external_url(request.url),
+        stored_state=session.pop('state', None),
+        received_state=request.args.get('state', ''))
+
+    session['credentials'] = credentials
+    session['user'] = user_info
+    current_app.logger.info(f"user_info={user_info}")
+
+    return redirect(session.pop('login_return', url_for('.list')))
+
+
+@app.route("/logout")
+def logout():
+    """
+    Log out and return to root page.
+    """
+    logout_session()
+    return redirect(url_for('.list'))
+
+
 @app.route('/')
 def list():
     books = booksdb.list()
@@ -43,6 +165,10 @@ def view(book_id):
 
 @app.route('/books/add', methods=['GET', 'POST'])
 def add():
+    # must be logged in
+    if "credentials" not in session:
+        session['login_return'] = url_for('.add')
+        return redirect(url_for('.login'))
     if request.method == 'POST':
         data = request.form.to_dict(flat=True)
 
@@ -58,6 +184,10 @@ def add():
 
 @app.route('/books/<book_id>/edit', methods=['GET', 'POST'])
 def edit(book_id):
+    # must be logged in
+    if "credentials" not in session:
+        session['login_return'] = url_for('.edit', book_id=book_id)
+        return redirect(url_for('.login'))
     book = booksdb.read(book_id)
     if request.method == 'POST':
         data = request.form.to_dict(flat=True)
@@ -74,8 +204,46 @@ def edit(book_id):
 
 @app.route('/books/<book_id>/delete')
 def delete(book_id):
+    # must be logged in
+    if "credentials" not in session:
+        session['login_return'] = url_for('.view', book_id=book_id)
+        return redirect(url_for('.login'))
     booksdb.delete(book_id)
     return redirect(url_for('.list'))
+
+
+@app.route('/profile', methods=['GET', 'POST'])
+def profile():
+    """
+    If GET, show the form to collect updated details for the user profile.
+    If POST, update the profile based on the specified form.
+    """
+    # must be logged in
+    if "credentials" not in session:
+        session['login_return'] = url_for('.profile')
+        return redirect(url_for('.login'))
+
+    # read existing profile
+    email = session['user']['email']
+    profile = profiledb.read(email)
+
+    # Save details if form was posted
+    if request.method == 'POST':
+
+        # get book details from form
+        data = request.form.to_dict(flat=True)
+
+        # update profile
+        profiledb.update(data, email)
+        session['preferred_language'] = data['preferredLanguage']
+
+        # return to root
+        return redirect(url_for('.list'))
+
+    # render form to update book
+    return render_template('profile.html', action='Edit',
+        profile=profile, languages=translate.get_languages())
+
 
 # this is only used when running locally
 if __name__ == '__main__':
